Starting From Scratch
Essentially starting from scratch, have a more focused direction in mind in terms of visual and placment of elements. Basic JS framework is more or less the same just with functions gutted and left as a skeleton for new logic. CSS is wiped clean from previous version, ditto for HTML.